requirements.txt
pillow
requests
bs4
lxml